aman

+++ model/url.go
@@ -9,7 +9,7 @@ import (
 type Url struct {
 	ID 				primitive.ObjectID 	`bson:"_id,omitempty" json:"id"` //
 	UserID 			primitive.ObjectID 	`bson:"user_id" json:"userid"` //
-	URLID			uint				`bson:"url_id" json:"url_id"`
+	URLID			int64				`bson:"url_id" json:"url_id"`
 	URL 			string 				`bson:"url" json:"url"` //
 	ShortLink 		string 				`bson:"shortlink" json:"shortlink"` //
 	ClickCount 		int 				`bson:"click_count" json:"clickcount"` //
